+++ admin/js/wdevs-tax-switch-woocommerce.js
@@ -1,4 +1,3 @@
-/* global jQuery, wp */
 ( function ( $ ) {
 	const colorDefaults = {
 		switch_color: '#333333',
@@ -9,12 +8,10 @@
 	};
 
 	$( function () {
-		// Basis shortcode bij load
 		$( '#wdevs-tax-switch-shortcode' ).val(
 			'[wdevs_tax_switch switch-type="switch" switch-color="#333333" switch-color-checked="#ffffff" switch-bg-color="#e9e9ea" switch-bg-color-checked="#34c759" switch-text-color="#333333" class-name="is-style-default"]'
 		);
 
-		// Modal openen
 		$( '#wdevs-generate-shortcode' ).on( 'click', function ( e ) {
 			e.preventDefault();
 			$( this ).WCBackboneModal( {
@@ -22,19 +19,18 @@
 			} );
 		} );
 
-		// Modal response afhandelen
 		$( document.body ).on(
 			'wc_backbone_modal_response',
-			function ( e, target, posted_data ) {
+			function ( e, target, postedData ) {
 				if ( target === 'wdevs-tax-switch-modal' ) {
 					let attributes = [];
 
 					// Directe mapping naar shortcode attributes
-					Object.keys( posted_data ).forEach( function ( key ) {
-						if ( posted_data[ key ] ) {
+					Object.keys( postedData ).forEach( function ( key ) {
+						if ( postedData[ key ] ) {
 							attributes.push(
 								`${ key.replace( /_/g, '-' ) }="${
-									posted_data[ key ]
+									postedData[ key ]
 								}"`
 							);
 						}
@@ -47,12 +43,10 @@
 			}
 		);
 
-		// Color pickers initialiseren
 		$( document ).on( 'wc_backbone_modal_loaded', function ( e, target ) {
 			if ( target === 'wdevs-tax-switch-modal' ) {
 				$( '.color-picker' ).wpColorPicker();
 
-				// Kleuren instellen
 				Object.keys( colorDefaults ).forEach( function ( key ) {
 					$( `input[name="${ key }"]` )
 						.val( colorDefaults[ key ] )