+++ includes/class-wdevs-tax-switch-block.php
@@ -11,36 +11,35 @@
  */
 
 /**
- * The block functionality of the plugin.
+ * The abstract block functionality of the plugin.
  *
- * Defines the plugin name, version, and hooks for the block functionality.
- * This class is responsible for registering and rendering the tax switch block and shortcode.
+ * Defines the structure and common functions of a Tax Switch block.
  *
  * @package    Wdevs_Tax_Switch
  * @subpackage Wdevs_Tax_Switch/includes
  * @author     Wijnberg Developments <contact@wijnberg.dev>
  */
-class Wdevs_Tax_Switch_Block {
+abstract class Wdevs_Tax_Switch_Block {
 
 	use Wdevs_Tax_Switch_Helper;
 
 	/**
 	 * The ID of this plugin.
 	 *
-	 * @since    1.0.0
+	 * @since    1.5.0
 	 * @access   private
 	 * @var      string $plugin_name The ID of this plugin.
 	 */
-	private $plugin_name;
+	protected $plugin_name;
 
 	/**
 	 * The version of this plugin.
 	 *
-	 * @since    1.0.0
+	 * @since    1.5.0
 	 * @access   private
 	 * @var      string $version The current version of this plugin.
 	 */
-	private $version;
+	protected $version;
 
 
 	/**
@@ -49,7 +48,7 @@ class Wdevs_Tax_Switch_Block {
 	 * @param string $plugin_name The name of this plugin.
 	 * @param string $version The version of this plugin.
 	 *
-	 * @since    1.0.0
+	 * @since    1.5.0
 	 */
 	public function __construct( $plugin_name, $version ) {
 
@@ -57,49 +56,21 @@ class Wdevs_Tax_Switch_Block {
 		$this->version     = $version;
 	}
 
-	public function init_block() {
-		register_block_type( plugin_dir_path( dirname( __FILE__ ) ) . 'build/block.json', array(
-			'render_callback' => [ $this, 'block_render_callback' ],
-		) );
 
-		register_block_style( 'wdevs/tax-switch', [
-			'name'  => 'inline',
-			'label' => __( 'Inline style', 'tax-switch-for-woocommerce' ),
-		] );
+	abstract public function register_frontend_scripts();
+	abstract public function init_block();
+	abstract public function register_shortcode();
 
-		wp_set_script_translations( 'wdevs-tax-switch-editor-script', 'tax-switch-for-woocommerce', plugin_dir_path( dirname( __FILE__ ) ) . 'languages' );
-	}
-
-	//https://developer.woocommerce.com/2021/11/15/how-does-woocommerce-blocks-render-interactive-blocks-in-the-frontend/
-	public function block_render_callback( $attributes = [], $content = '' ) {
-		if ( ! is_admin() ) {
-			$this->enqueue_frontend_scripts();
-		}
-
-		return $this->add_attributes_to_block( $attributes, $content );
-	}
-
-	public function register_shortcode() {
-		add_shortcode( 'wdevs_tax_switch', array( $this, 'shortcode_render_callback' ) );
-	}
-
-	public function shortcode_render_callback( $attributes = [], $content = '' ) {
-		if ( ! is_admin() ) {
-			$this->enqueue_frontend_scripts();
-		}
-
-		$attributes = shortcode_atts( [
-			'class-name'                      => 'is-style-default',
-			'switch-type'                     => 'switch',
-			'switch-color'                    => '',
-			'switch-color-checked'            => '',
-			'switch-background-color'         => '',
-			'switch-background-color-checked' => '',
-			'switch-text-color'               => '',
-			'switch-label-incl'               => '',
-			'switch-label-excl'               => '',
-		], $attributes );
-
+	/**
+	 * @param array $attributes
+	 * @param string $container_class_name
+	 * @param string $content
+	 *
+	 * @return string
+	 *
+	 * @since 1.5.0
+	 */
+	protected function render_shortcode_html( array $attributes, string $container_class_name, string $content, string $element = 'div' ): string {
 		$wpml_active             = defined( 'ICL_SITEPRESS_VERSION' );
 		$translatable_attributes = [ 'switch-label-incl', 'switch-label-excl' ];
 		foreach ( $translatable_attributes as $label_key ) {
@@ -111,63 +82,25 @@ class Wdevs_Tax_Switch_Block {
 			}
 		}
 
-		$holder_class_name = 'wp-block-wdevs-tax-switch'; //important for rendering JS
+
 		if ( isset( $attributes['class-name'] ) && ! empty( $attributes['class-name'] ) ) {
-			$holder_class_name .= ' ' . esc_attr($attributes['class-name']);
+			$container_class_name .= ' ' . esc_attr( $attributes['class-name'] );
 		}
 
-		$content = '<div class="' . $holder_class_name . '"></div>';
+		$content = '<div class="' . $container_class_name . '"></div>';
 
 		return $this->add_attributes_to_block( $attributes, $content );
 	}
 
 	/**
-	 * @since 1.2.4
+	 * @param array $attributes
+	 * @param string $content
+	 *
+	 * @return string
+	 *
+	 * @since 1.5.0
 	 */
-	public function register_frontend_scripts() {
-		$script_asset = require( plugin_dir_path( dirname( __FILE__ ) ) . 'build/view.asset.php' );
-
-		wp_register_script(
-			'wdevs-tax-switch-view-script',
-			plugin_dir_url( dirname( __FILE__ ) ) . 'build/view.js',
-			$script_asset['dependencies'],
-			$script_asset['version']
-		);
-
-		wp_register_style(
-			'wdevs-tax-switch-style',
-			plugin_dir_url( dirname( __FILE__ ) ) . 'build/style-index.css',
-			[],
-			$script_asset['version']
-		);
-	}
-
-	public function enqueue_frontend_scripts() {
-		if ( wp_style_is( 'wdevs-tax-switch-style', 'registered' ) ) {
-			wp_enqueue_style( 'wdevs-tax-switch-style' );
-		}
-
-		if ( wp_script_is( 'wdevs-tax-switch-view-script', 'registered' ) ) {
-			wp_enqueue_script( 'wdevs-tax-switch-view-script' );
-
-			$original_tax_display = $this->get_original_tax_display();
-			wp_localize_script(
-				'wdevs-tax-switch-view-script',
-				'wtsViewObject',
-				[
-					'originalTaxDisplay' => $original_tax_display
-				]
-			);
-
-			wp_set_script_translations(
-				'wdevs-tax-switch-view-script',
-				'tax-switch-for-woocommerce',
-				plugin_dir_path( dirname( __FILE__ ) ) . 'languages'
-			);
-		}
-	}
-
-	public function add_attributes_to_block( $attributes = [], $content = '' ) {
+	protected function add_attributes_to_block( $attributes = [], $content = '' ) {
 		$escaped_data_attributes = [];
 
 		foreach ( $attributes as $key => $value ) {