+++ webpack.config.js
@@ -4,37 +4,55 @@ const path = require( 'path' );
 module.exports = {
 	...defaultConfig,
 	entry: {
-		index: path.resolve( process.cwd(), 'block/src', 'index.js' ),
-		view: path.resolve( process.cwd(), 'block/src', 'view.js' ),
-		'woocommerce-tiered-price-table': path.resolve(
+		'switch/index': path.resolve(
 			process.cwd(),
-			'block/src',
+			'block/src/switch',
+			'index.js'
+		),
+		'switch/view': path.resolve(
+			process.cwd(),
+			'block/src/switch',
+			'view.js'
+		),
+		'switch/woocommerce-tiered-price-table': path.resolve(
+			process.cwd(),
+			'block/src/switch',
 			'woocommerce-tiered-price-table.js'
 		),
-		'woocommerce-measurement-price-calculator': path.resolve(
+		'switch/woocommerce-measurement-price-calculator': path.resolve(
 			process.cwd(),
-			'block/src',
+			'block/src/switch',
 			'woocommerce-measurement-price-calculator.js'
 		),
-		'yith-woocommerce-product-add-ons': path.resolve(
+		'switch/yith-woocommerce-product-add-ons': path.resolve(
 			process.cwd(),
-			'block/src',
+			'block/src/switch',
 			'yith-woocommerce-product-add-ons.js'
 		),
-		'woocommerce-product-addons': path.resolve(
+		'switch/woocommerce-product-addons': path.resolve(
 			process.cwd(),
-			'block/src',
+			'block/src/switch',
 			'woocommerce-product-addons.js'
 		),
-		'advanced-product-fields-for-woocommerce': path.resolve(
+		'switch/advanced-product-fields-for-woocommerce': path.resolve(
 			process.cwd(),
-			'block/src',
+			'block/src/switch',
 			'advanced-product-fields-for-woocommerce.js'
 		),
-		'woocommerce-quantity-manager': path.resolve(
+		'switch/woocommerce-quantity-manager': path.resolve(
 			process.cwd(),
-			'block/src',
+			'block/src/switch',
 			'woocommerce-quantity-manager.js'
 		),
-	},
+		'label/index': path.resolve(
+			process.cwd(),
+			'block/src/label',
+			'index.js'
+		),
+		'label/view': path.resolve(
+			process.cwd(),
+			'block/src/label',
+			'view.js'
+		),
+	}
 };